Create Youtube File Uploader.py
This is the PY file construct. To use this properly, you will need to sign up for a google developers account and create a project that uses the YouTube API. You get 50,000,000 free interactions per day. Unless you are looking to do a huge amount of file transmissions, this should be plenty.

More information can be found in the README to come.